# backend/api-docs.yaml
openapi: 3.0.0
info:
  title: YouTube Clone API
  version: 1.0.0
  description: |
    Complete API documentation for YouTube Clone application.
    This API provides endpoints for user management, video uploads, and comments.
  contact:
    name: API Support
    email: support@youtubeclone.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.youtubeclone.com
    description: Production server
...
paths:
...
  /api/v1/comment/new:
    post:
      tags:
        - Comments
      summary: Add a new comment
      description: Add a comment to a video. Requires authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - video_id
                - commentText
              properties:
                video_id:
                  type: string
                  example: "507f1f77bcf86cd799439011"
                commentText:
                  type: string
                  example: "Great video!"
      responses:
        '201':
          description: Comment added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment added successfully"
                  comment:
                    $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/comment/comment/{videoId}:
    get:
      tags:
        - Comments
      summary: Get comments for a video
      description: Retrieve all comments for a specific video. No authentication required.
      parameters:
        - in: path
          name: videoId
          required: true
          schema:
            type: string
          description: The video ID
          example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/comment/{commentId}:
    put:
      tags:
        - Comments
      summary: Update a comment
      description: Update a comment. Only the comment owner can update.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: commentId
          required: true
          schema:
            type: string
          description: Comment ID
          example: "507f1f77bcf86cd799439011"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - commentText
              properties:
                commentText:
                  type: string
                  example: "Updated comment text"
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment updated successfully"
                  comment:
                    $ref: '#/components/schemas/Comment'
        '403':
          description: Unauthorized - not the comment owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Comments
      summary: Delete a comment
      description: Delete a comment. Only the comment owner can delete.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: commentId
          required: true
          schema:
            type: string
          description: Comment ID
          example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Comment deleted successfully"
        '403':
          description: Unauthorized - not the comment owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        ChannelName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        logoUrl:
          type: string
          format: uri
        subscriber:
          type: number
          default: 0
        subscribedChannels:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
...
    Comment:
      type: object
      properties:
        _id:
          type: string
        video_id:
          type: string
        user_id:
          type: string
        commentText:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...